added tts, file opening, copying, and moving
She now speaks, and has improved language parsing. Can recognize what part of a request is referring to a file/directory, and make inferences if she doesn't immediately a set of words as a file. Still needs improvements on that last part though. Also  redistributed the weights for the word vectors that represent words that would be contained within the name of a file.